production mode 再delete device

+++ notifier/notifier.py
@@ -254,6 +254,7 @@ class NotifierService(object):
         msg = self.composeAPNSMessage(alert, badge)
         self.logger.info("start to send notification for user: " + notifier.user_id)
         device_list = notifier.device_id
+        self.logger.info("device list:" + str(device_list))
         invalid_device = []
         isSend = False
         for device in device_list:
@@ -270,7 +271,7 @@ class NotifierService(object):
                 if isSend == False:
                     self.logger.info("Cannot find any device to send for user: " + notifier.user_id)
 
-                if len(invalid_device) > 0:
+                if len(invalid_device) > 0 and LocalConstant.PRODUCT_MODE == True:
                     self.notifierWeb.deleteDevices(notifier.user_id, invalid_device)
             except:
                 pass
